requirements.txt
# Core dependencies
anthropic>=0.40.0
slack-bolt>=1.18.0
slack-sdk>=3.27.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0
